LFOPTEE-241 drivers: ele: add ELE status code and debug traces
Add all HSM and ELE status codes.
Add debug trace for HSM and ELE MU calls.
