+++ core/drivers/crypto/ele/sub.mk
@@ -4,4 +4,5 @@ srcs-y += ele.c
 srcs-y += utils_mem.c
 srcs-y += key_store.c
 srcs-y += key_mgmt.c
+srcs-y += utils_trace.c
 subdirs-$(CFG_IMX_ELE_ACIPHER_DRV) += acipher
